Chapitre 77 : Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -64,10 +64,13 @@ router.put('/:id', checkAuth, multer({storage: storage}).single('image'), (req, 
     imagePath: imagePath
   });
   console.log(post);
-  Post.updateOne({_id: req.params.id}, post)
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post)
     .then(result => {
-
+      if (result.modifiedCount > 0) {
       res.status(200).json({ message: "update successful"});
+      } else {
+        res.status(401).json({message: "Not authorized"});
+      }
     });
 });
 
@@ -106,10 +109,13 @@ router.get('/:id',(req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
     .then(result => {
-      console.log (result);
-      res.status(200).json({ message: "Post deleted!" });
+      if (result.deletedCount > 0) {
+        res.status(200).json({ message: "Post deleted!" });
+      } else {
+        res.status(401).json({ message: "Not authorized!" })
+      }
     });
 });
